Upload breakpad minidumps from an android service
